requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-SocketIO==5.3.3
Flask-Login==0.6.2
websocket-client==1.5.1
twilio==7.16.0
python-dotenv==0.21.0
